Source code cleanup
